ref(kernel): zram splits out of swap.py
* generator config and package on one side, the hibernation swap file and
  its resume params on the other
